GUI-258. Profiles: move profile types to common
