+++ front/src/common/types/index.ts
@@ -1,0 +1,4 @@
+export interface IListResult<T> {
+    records: Array<T>;
+    total: number;
+}
