+++ front/src/common/types/profile-attribute.ts
@@ -1,0 +1,4 @@
+export interface IAttribute {
+    label: string;
+    value: string;
+}
