+++ front/src/common/types/profile-certificate.ts
@@ -1,0 +1,6 @@
+import { EnumProfileStatus } from './profile-status';
+
+export interface ICertificate {
+    status: EnumProfileStatus;
+    address: string;
+}
