+++ front/src/common/types/profile-status.ts
@@ -1,0 +1,7 @@
+export enum EnumProfileStatus {
+    anonimest = 0,
+    anon = 1,
+    reg = 2,
+    ident = 3,
+    pro = 4,
+}
